validation added

+++ twitter_post.py
@@ -25,22 +25,22 @@ class TwitterPost(TwitterProp):
 		return False
 
 	def _post(self, post, file_path):
-		if len(post) < 250:
-			textbox = self.page.locator(global_config["post_textarea_selector"])
-			textbox.type(post)
-			textbox.type(" ")
+		textbox = self.page.locator(global_config["post_textarea_selector"])
+		textbox.type(post)
+		textbox.type(" ")
 
-			file_input = self.page.locator('input[type="file"]')
-			file_input.set_input_files(file_path)
+		file_input = self.page.locator('input[type="file"]')
+		file_input.set_input_files(file_path)
 
-			button_locator = self.page.locator(global_config["post_tweet_selector"])
-			button_locator.wait_for(state="visible", timeout=50000)
+		button_locator = self.page.locator(global_config["post_tweet_selector"])
+		button_locator.wait_for(state="visible", timeout=50000)
 
-			self.page.wait_for_timeout(2000)
-			# while button_locator.is_disabled():
-			# 	self.page.wait_for_timeout(100)
+		count = 50
+		while button_locator.is_disabled() and count > 0:
+			count -= 1
+			self.page.wait_for_timeout(1000)
 
-			x_utils.click(self.page, global_config["post_tweet_selector"])
+		x_utils.click(self.page, global_config["post_tweet_selector"])
 
 	def start(self):
 		if global_config["media_path"]:
@@ -53,11 +53,13 @@ class TwitterPost(TwitterProp):
 					break
 
 				file_path = random.choice([file for file in common.list_files_recursive(global_config["media_path"]) if file.endswith(".png") or file.endswith(".jpg") or file.endswith(".mkv") or file.endswith(".mp4")])
+				file_path  = "/home/jebineinstein/git/CaptionCreator/media/testing/test_image.png"
 
 				count += 1
 				geminiWrapper = GeminiWrapper(system_instruction=global_config["post_sp"])
 				model_responses = geminiWrapper.send_message("", file_path=file_path)
 				response = json.loads(model_responses[0])
-				self._post(response["post"], file_path)
-				if global_config["delete_media_path_after_post"] == "yes":
-					common.remove_file(file_path)
+				if response["can_post"] == "yes" and len(response["post"]) < 250:
+					self._post(response["post"], file_path)
+					if global_config["delete_media_path_after_post"] == "yes":
+						common.remove_file(file_path)
